Removed EOL Versions and added Composer
First, all PHPUnit EOL Versions (5,6) are deleted. It's not an
option to use such old versions.

Second, the Base-Image has changed to the pimlab/composer image,
because its often usefull to install the vendor for your tests.
But if you test on non deploy pipelines, it is not an good idea
to upload artifacts like you do on deployment.